refactor: move loans type to types folder
