+++ src/types/loans.ts
